backend/pyproject.toml
[project]
name = "polydata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-oidc-provider>=0.8.4",
    "django-unfold>=0.74.1",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "geopandas>=1.1.3",
    "gunicorn>=25.1.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.48",
    "wagtail>=7.3.1",
    "whitenoise>=6.12.0",
    "pyyaml>=6.0",
    "pygeoapi>=0.23.0",
    "geoalchemy2>=0.18.4",
]

[tool.ruff]
fix = true
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]
exclude = ["/*/migrations/**"]
